feat: Implementando esquema de autorização com gates e policies.

+++ app/Http/Controllers/BrandController.php
@@ -8,6 +8,7 @@ use App\Models\Brand;
 use App\Services\BrandServices;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class BrandController extends Controller
 {
@@ -21,6 +22,7 @@ class BrandController extends Controller
      */
     public function index(): JsonResponse
     {
+        Gate::authorize('viewAny', Brand::class);
         return response()->json($this->brandServices->list());
     }
 
@@ -29,6 +31,7 @@ class BrandController extends Controller
      */
     public function store(BrandStoreRequest $request): JsonResponse
     {
+        Gate::authorize('create', Brand::class);
         return response()->json($this->brandServices->store($request));
     }
 
@@ -37,6 +40,7 @@ class BrandController extends Controller
      */
     public function show(Brand $brand): JsonResponse
     {
+        Gate::authorize('view', $brand);
         return response()->json($brand);
     }
 
@@ -45,6 +49,7 @@ class BrandController extends Controller
      */
     public function update(BrandUpdateRequest $request, Brand $brand): JsonResponse
     {
+        Gate::authorize('update', $brand);
         return response()->json($this->brandServices->update($request, $brand));
     }
 
@@ -53,6 +58,7 @@ class BrandController extends Controller
      */
     public function destroy(Brand $brand)
     {
+        Gate::authorize('delete', $brand);
         return response()->json($this->brandServices->delete($brand));
     }
 }
